test (ResourcesResponderV2): Add debugging code.
